Remove a validation rule
A validation rule was checking the UI state which shouldn't be done in the model (the UI worked anyway). It also gave a false negative when opening really simple saved puzzles where the value of a cell was able to be calculated before it was later set by the files contents.