Add a simple trial and error strategy
Occasionally a puzzle will have 4 cells at the intersection of 2 rows and 2 columns which can have 2 values but the combination of the values is in determinant, either is valid. This new strategy attempts to solve the puzzle by inserting a possible value into one of the cells and seeing if the puzzle can then be solved. As it turns out this strategy also allows previously unsolvable puzzles to be solved. This may be the strategy that I was missing all along.