Add an Enum mapper generic class (Mk. 2)
Add an enum mapper class that is used to display debug info but is also intended to be used when saving puzzles where the origins property will  be added to the xml as a cell attribute.
(this time use a lock to protect the mapper initialisation, d'oh! - not enough caffeine)