Allow calculated cells to be edited
The trouble with solving as you type is that the puzzle can be solved before you've typed in all the numbers. While the solution is valid it isn't the same as puzzle that was intended. This change allows calculated cell values to be edited so the puzzle can be entered correctly.