Rewrite the SudokuGrid
The old code overrode the OnRender() method and drew the grid lines on top of the panel. This wouldn't work on UWP because drawing is further abstracted and there is no OnRender() method to override. The code now specifies the grid lines in xaml and uses the Measure/Arrange layout passes to set the line dimensions.