Fix opening saved puzzles when hiding the solution
When opening a file the code ignored any user entered values if it had already calculated the cells value. This was ok before you could hide the solution, but now user entered cells always have to be identical to the saved puzzle. The code will now only recalculate the puzzle once when its finished reading the file rather than after reading each cell (even if that was useful for probably excessive file validation).